// src/api/openapi/auto-generated/models/ResultSetResponse.ts
export interface ResultSetResponse {
    [key: string]: any | any;
    /**
     * 
     * @type {Options5}
     * @memberof ResultSetResponse
     */
    options: Options5;
    /**
     * 
     * @type {string}
     * @memberof ResultSetResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ResultSetResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ResultSetResponse
     */
    portfolio: string;
    /**
     * 
     * @type {ResultSetStatus}
     * @memberof ResultSetResponse
     */
    status: ResultSetStatus;
    /**
     * 
     * @type {Date}
     * @memberof ResultSetResponse
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof ResultSetResponse
     */
    updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ResultSetResponse
     */
    createdBy: string;
    /**
     * 
     * @type {string}
     * @memberof ResultSetResponse
     */
    updatedBy: string;
    /**
     * 
     * @type {ResultSetResponseStatistics}
     * @memberof ResultSetResponse
     */
    statistics: ResultSetResponseStatistics;
}